requirements.txt
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.14.0
Flask==3.0.3
fsspec==2024.5.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai-whisper==20231117
regex==2024.5.15
requests==2.32.3
scipy==1.13.1
sympy==1.12.1
tiktoken==0.7.0
torch==2.3.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
